Changed Nonce vairable type and Updated createGenesis function

+++ src/main/java/co/usc/ulordj/params/RegTestParams.java
@@ -55,7 +55,7 @@ public class RegTestParams extends TestNet2Params {
         synchronized (RegTestParams.class) {
             if (genesis == null) {
                 genesis = super.getGenesisBlock();
-                genesis.setNonce(2);
+                genesis.setNonce(BigInteger.valueOf(2));
                 genesis.setDifficultyTarget(0x207fFFFFL);
                 genesis.setTime(1296688602L);
                 checkState(genesis.getHashAsString().toLowerCase().equals("0f9188f13cb7b2c71f2a335e3a4fc328bf5beb436012afca590b1a11466e2206"));
